requirements.txt
flask
flask_cors
openai
python-dotenv
beautifulsoup4
scipy